move main interface tests to folder
